except malloc-almost everything

+++ include/string.h
@@ -2,7 +2,7 @@
 #define _STRING_H
 
 int strlen(const char *s);
-int strcmp(char *s1, char *s2,uint64_t n);
+int strcmp(char *s1, char *s2,int n);
 int strcmp1(char *s1, char *s2);
 int first_occurence(char *s,char c);
 char *strtok(char *str,char *delim, char* token);
